Fillters and Graph
A filther for the table wich now you can choose a letter and the information in the table will be filtered so only the departments which start with the selected letter will be shown.

makes a cicular graph with the information that has been loaded into the program.

some design changes has been made in order to display better the information and be more gentle to a final user